Minor Update to Example Scene
Moves Distance Curve Noise Objects into a separate class.
