+++ Runtime/DistanceCurveNoiseObject.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ffecfce54e349284da4164addfabf2a4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
