+++ Runtime/NoiseObjects/New Curve Noise.asset
@@ -9,14 +9,41 @@ MonoBehaviour:
   m_GameObject: {fileID: 0}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 0}
+  m_Script: {fileID: 11500000, guid: ffecfce54e349284da4164addfabf2a4, type: 3}
   m_Name: New Curve Noise
-  m_EditorClassIdentifier: quiqi.knightgen.Runtime::DistanceCurveNoiseObject
+  m_EditorClassIdentifier: 
   scale: 2
   max: 1
   curve:
     serializedVersion: 2
-    m_Curve: []
+    m_Curve:
+    - serializedVersion: 3
+      time: 0
+      value: 0
+      inSlope: 0
+      outSlope: 0
+      tangentMode: 0
+      weightedMode: 0
+      inWeight: 0
+      outWeight: 0
+    - serializedVersion: 3
+      time: 0.6836526
+      value: 0.39448333
+      inSlope: -0.37686124
+      outSlope: -0.37686124
+      tangentMode: 0
+      weightedMode: 0
+      inWeight: 0.25644496
+      outWeight: 0.33333334
+    - serializedVersion: 3
+      time: 1
+      value: 1
+      inSlope: 0.0038188251
+      outSlope: 0.0038188251
+      tangentMode: 0
+      weightedMode: 0
+      inWeight: 1
+      outWeight: 0
     m_PreInfinity: 2
-    m_PostInfinity: 2
+    m_PostInfinity: 0
     m_RotationOrder: 4
